[bioshd] Fix floppy multitrack issue, merge raw IO into driver
